+++ src/objects/statchange.h
@@ -196,7 +196,7 @@ const std::unordered_map<std::string, _StatStorage> StatStringToType = {
 	{ "BleedPower",             { StatType::BLEEDPOWER             , StatValueType::FLOAT   } },
 	{ "PoisonPower",            { StatType::POISONPOWER            , StatValueType::FLOAT   } },
 	{ "PetPower",               { StatType::PETPOWER               , StatValueType::FLOAT   } },
-	{ "HealPower",              { StatType::HEALPOWER              , StatValueType::FLOAT   } },
+	{ "HealPower",              { StatType::HEALPOWER              , StatValueType::INTEGER } },
 	{ "ManaPower",              { StatType::MANAPOWER              , StatValueType::INTEGER } },
 	{ "Respec",                 { StatType::RESPEC                 , StatValueType::BOOLEAN } },
 	{ "Corpse",                 { StatType::CORPSE                 , StatValueType::INTEGER } },
